+++ lib/gembrew/commands/build.rb
@@ -4,14 +4,25 @@ require 'gembrew/generator'
 module Gembrew
   module Commands
     class Build < MisterBin::Command
-      help 'Generate a Homebrew formula from gembrew.yml'
+      help 'Generate Homebrew formulae from gembrew/*.yml'
 
-      usage 'gembrew build'
+      usage 'gembrew build [GEM]'
       usage 'gembrew build (-h|--help)'
 
+      param 'GEM', 'Optional gem configuration name'
+
       def run
-        output_path = Generator.new.call
-        say "Generated m`#{output_path}`"
+        configs.each do |config|
+          say "bb`==> gembrew: building #{config.name}`"
+          output_path = Generator.new(config: config).call
+          say "Generated m`#{output_path}`"
+        end
+      end
+
+    private
+
+      def configs
+        args['GEM'] ? [Config.find(args['GEM'])] : Config.all
       end
     end
   end