+++ lib/gembrew/commands/check.rb
@@ -5,16 +5,27 @@ require 'gembrew/generator'
 module Gembrew
   module Commands
     class Check < MisterBin::Command
-      help 'Build and test the formula in a clean Homebrew environment'
+      help 'Build and test formulae in clean Homebrew environments'
 
-      usage 'gembrew check'
+      usage 'gembrew check [GEM]'
       usage 'gembrew check (-h|--help)'
 
+      param 'GEM', 'Optional gem configuration name'
+
       def run
-        output_path = Generator.new.call
-        say "Generated m`#{output_path}`"
-        DockerCheck.new.call
+        configs.each do |config|
+          say "bb`==> gembrew: building #{config.name}`"
+          output_path = Generator.new(config: config).call
+          say "Generated m`#{output_path}`"
+          DockerCheck.new.call output_path.basename('.rb').to_s
+        end
         say 'g`All checks passed`'
+      end
+
+    private
+
+      def configs
+        args['GEM'] ? [Config.find(args['GEM'])] : Config.all
       end
     end
   end